+++ ios/Classes/Unified/AdsterUnifiedAd.swift
@@ -1,4 +1,5 @@
 import Flutter
+import SafariServices
 import AdsFramework
 
 class AdsterUnifiedAd : NSObject{
@@ -57,7 +58,23 @@ class AdsterUnifiedAd : NSObject{
     }
     
     func click(compname: String) {
-        self.mediaViewByAd?.gestureRecognizers?.first?.state = .ended
+        guard let nativeAd = self.nativeAdView else { return }
+
+        guard let landingUrl = nativeAd.landingUrl,
+              let url = URL(string: landingUrl) else {
+            return
+        }
+
+        nativeAd.recordNativeClick()
+        
+        DispatchQueue.main.async {
+            if nativeAd.openUrlInApp {
+                let safariVC = SFSafariViewController(url: url)
+                UIApplication.shared.windows.first?.rootViewController?.present(safariVC, animated: true)
+            } else if UIApplication.shared.canOpenURL(url) {
+                UIApplication.shared.open(url, options: [:], completionHandler: nil)
+            }
+        }
     }
 }
 
@@ -102,9 +119,11 @@ extension AdsterUnifiedAd: MediationAdDelegate {
 
     func onNativeAdLoaded(nativeAd: any AdsFramework.MediationNativeAd) {
         self.nativeAdView = nativeAd
-        self.nativeAdView?.eventCallbacks = adsterRevenueOnlyCallbacks(
+        self.nativeAdView?.eventCallbacks = adsterEventCallbacks(
             widgetId: widgetId,
-            channel: adClickChannel
+            channel: adClickChannel,
+            clickMethod: "onAdClicked",
+            impressionMethod: "onAdImpression"
         )
         self.onNativeAdLoadComplete?(widgetId,nativeAd)
     }